pmfs working -- but slow?

+++ gsl_server/include/gsl_server/algorithms/PMFS/internal/VisibilityMap.hpp
@@ -9,7 +9,7 @@ namespace GSL
     class VisibilityMap
     {
     public:
-        size_t range = 0;
+        uint range = 0;
 
         Visibility isVisible(const Vector2Int& from, const Vector2Int& to) const
         {
